Refactor utils functions and constants into utils module
