+++ app/build.gradle
@@ -118,4 +118,6 @@ dependencies {
     coreLibraryDesugaring 'com.android.tools:desugar_jdk_libs:1.2.0'
 
     implementation(project(":core:ui"))
+    implementation(project(":core:translations")) // RECHECK
+    implementation(project(":core:util"))
 }
