+++ app/src/main/java/com/xdiach/diarymoodapp/data/repository/MongoDB.kt
@@ -1,7 +1,7 @@
 package com.xdiach.diarymoodapp.data.repository
 
-import com.xdiach.diarymoodapp.model.Diary
-import com.xdiach.diarymoodapp.model.RequestState
+import com.xdiach.util.model.Diary
+import com.xdiach.util.model.RequestState
 import com.xdiach.diarymoodapp.util.Constants.APP_ID
 import com.xdiach.diarymoodapp.util.toInstant
 import io.realm.kotlin.Realm
