+++ app/src/main/java/com/xdiach/diarymoodapp/di/DatabaseModule.kt
@@ -2,7 +2,7 @@ package com.xdiach.diarymoodapp.di
 
 import android.content.Context
 import androidx.room.Room
-import com.xdiach.diarymoodapp.connectivity.NetworkConnectivityObserver
+import com.xdiach.util.connectivity.NetworkConnectivityObserver
 import com.xdiach.diarymoodapp.data.database.ImagesDatabase
 import com.xdiach.diarymoodapp.util.Constants.IMAGES_DATABASE
 import dagger.Module
