+++ app/src/main/java/com/xdiach/diarymoodapp/presentation/screens/home/HomeScreen.kt
@@ -21,7 +21,7 @@ import androidx.compose.ui.unit.LayoutDirection
 import androidx.compose.ui.unit.dp
 import com.xdiach.diarymoodapp.R
 import com.xdiach.diarymoodapp.data.repository.Diaries
-import com.xdiach.diarymoodapp.model.RequestState
+import com.xdiach.util.model.RequestState
 import java.time.ZonedDateTime
 
 @OptIn(ExperimentalMaterial3Api::class)
