+++ app/src/main/java/com/xdiach/diarymoodapp/presentation/screens/write/WriteScreen.kt
@@ -37,10 +37,10 @@ import coil.request.ImageRequest
 import com.google.accompanist.pager.ExperimentalPagerApi
 import com.google.accompanist.pager.PagerState
 import com.xdiach.diarymoodapp.R
-import com.xdiach.diarymoodapp.model.Diary
+import com.xdiach.util.model.Diary
 import com.xdiach.diarymoodapp.model.GalleryImage
 import com.xdiach.diarymoodapp.model.GalleryState
-import com.xdiach.diarymoodapp.model.Mood
+import com.xdiach.util.model.Mood
 import java.time.ZonedDateTime
 
 @OptIn(ExperimentalPagerApi::class)
