+++ core/ui/src/main/java/com/xdiach/ui/components/DiaryHolder.kt
@@ -45,14 +45,14 @@ import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.xdiach.translations.R
-import com.xdiach.diarymoodapp.model.Diary
-import com.xdiach.diarymoodapp.model.Mood
-import com.xdiach.diarymoodapp.util.fetchImagesFromFirebase
-import com.xdiach.diarymoodapp.util.toInstant
 import com.xdiach.ui.UiText
 import com.xdiach.ui.theme.DiaryMoodAppTheme
 import com.xdiach.ui.values.Dimensions
 import com.xdiach.ui.values.Elevation
+import com.xdiach.util.fetchImagesFromFirebase
+import com.xdiach.util.model.Diary
+import com.xdiach.util.model.Mood
+import com.xdiach.util.toInstant
 import java.text.DateFormat
 import java.time.Instant
 import java.util.Date
