+++ core/util/src/main/res/drawable/angry.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="utf-8"?>
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="36dp"
+    android:height="36dp"
+    android:viewportWidth="36"
+    android:viewportHeight="36">
+    <path
+        android:fillColor="#DA2F47"
+        android:pathData="M36,18C36,27.941 27.941,36 18,36C8.06,36 0,27.941 0,18C0,8.06 8.06,0 18,0C27.941,0 36,8.06 36,18Z" />
+    <path
+        android:fillColor="#292F33"
+        android:pathData="M25.485,29.879C25.44,29.7 24.317,25.5 18,25.5C11.682,25.5 10.56,29.7 10.515,29.879C10.46,30.096 10.558,30.321 10.752,30.433C10.947,30.542 11.191,30.512 11.352,30.356C11.371,30.337 13.306,28.5 18,28.5C22.694,28.5 24.63,30.337 24.648,30.355C24.744,30.45 24.872,30.5 25,30.5C25.084,30.5 25.169,30.479 25.246,30.436C25.442,30.324 25.54,30.097 25.485,29.879ZM15.707,17.293C12.452,14.038 7.221,14 7,14C6.448,14 6.001,14.447 6.001,14.998C6,15.55 6.447,15.998 6.999,16C7.028,16 8.924,16.022 10.982,16.737C10.389,17.377 10,18.371 10,19.5C10,21.434 11.119,23 12.5,23C13.881,23 15,21.434 15,19.5C15,19.326 14.981,19.16 14.963,18.993C14.976,18.993 14.988,19 15,19C15.256,19 15.512,18.902 15.707,18.707C16.098,18.316 16.098,17.684 15.707,17.293ZM29,14C28.779,14 23.549,14.038 20.293,17.293C19.902,17.684 19.902,18.316 20.293,18.707C20.488,18.902 20.744,19 21,19C21.013,19 21.024,18.993 21.036,18.993C21.02,19.16 21,19.326 21,19.5C21,21.434 22.119,23 23.5,23C24.881,23 26,21.434 26,19.5C26,18.371 25.611,17.377 25.018,16.737C27.076,16.022 28.972,16 29.002,16C29.553,15.998 30,15.55 29.999,14.998C29.998,14.447 29.552,14 29,14Z" />
+</vector>
