+++ core/util/src/main/res/drawable/awful.xml
@@ -1,0 +1,28 @@
+<?xml version="1.0" encoding="utf-8"?>
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="34dp"
+    android:height="36dp"
+    android:viewportWidth="34"
+    android:viewportHeight="36">
+    <path
+        android:fillColor="#FFCB4C"
+        android:pathData="M34,17C34,26.389 26.389,34 17,34C7.612,34 0,26.389 0,17C0,7.612 7.612,0 17,0C26.389,0 34,7.612 34,17Z" />
+    <path
+        android:fillColor="#65471B"
+        android:pathData="M17,27.136C20.335,27.136 22.547,27.524 25.284,28.056C25.909,28.176 27.125,28.056 27.125,26.215C27.125,22.533 22.895,17.931 17,17.931C11.105,17.931 6.875,22.533 6.875,26.215C6.875,28.056 8.091,28.177 8.716,28.056C11.452,27.524 13.665,27.136 17,27.136ZM5.668,18.888C5.381,18.888 5.097,18.758 4.912,18.51C4.599,18.093 4.683,17.501 5.101,17.188C5.236,17.087 7.888,15.127 11.576,14.418C9.238,13.246 6.644,13.221 6.611,13.221C6.09,13.22 5.668,12.796 5.669,12.275C5.669,11.754 6.091,11.332 6.613,11.332C6.822,11.332 11.762,11.368 14.836,14.442C15.106,14.712 15.187,15.118 15.041,15.471C14.895,15.824 14.55,16.054 14.169,16.054C9.812,16.054 6.271,18.672 6.236,18.698C6.063,18.827 5.865,18.888 5.668,18.888ZM28.332,18.888C28.136,18.888 27.937,18.827 27.768,18.701C27.732,18.674 24.168,16.056 19.833,16.056C19.451,16.056 19.106,15.826 18.961,15.473C18.814,15.12 18.895,14.714 19.166,14.444C22.241,11.369 27.181,11.334 27.389,11.334C27.91,11.334 28.332,11.756 28.333,12.277C28.334,12.798 27.911,13.221 27.391,13.223C27.357,13.223 24.762,13.248 22.425,14.42C26.113,15.128 28.765,17.088 28.9,17.19C29.317,17.503 29.402,18.093 29.09,18.511C28.904,18.758 28.62,18.888 28.332,18.888Z" />
+    <path
+        android:fillColor="#77AF57"
+        android:pathData="M24.712,32.287V25.186C24.712,22.593 20.823,20.001 16.934,20.001C13.045,20.001 9.156,22.594 9.156,25.186V32.287C4.753,32.636 1.797,33.272 1.797,34C1.797,35.105 8.574,36 16.934,36C25.294,36 32.071,35.105 32.071,34C32.071,33.272 29.115,32.636 24.712,32.287Z" />
+    <path
+        android:fillColor="#5D8F3F"
+        android:pathData="M14.875,24.463C15.557,24.463 16.11,23.91 16.11,23.228C16.11,22.546 15.557,21.993 14.875,21.993C14.193,21.993 13.64,22.546 13.64,23.228C13.64,23.91 14.193,24.463 14.875,24.463Z" />
+    <path
+        android:fillColor="#5D8F3F"
+        android:pathData="M19.698,26.875C20.38,26.875 20.933,26.322 20.933,25.64C20.933,24.958 20.38,24.405 19.698,24.405C19.016,24.405 18.463,24.958 18.463,25.64C18.463,26.322 19.016,26.875 19.698,26.875Z" />
+    <path
+        android:fillColor="#5D8F3F"
+        android:pathData="M21.452,30.712C21.561,30.039 21.104,29.404 20.431,29.295C19.758,29.186 19.123,29.643 19.014,30.316C18.905,30.99 19.362,31.624 20.035,31.733C20.708,31.843 21.343,31.385 21.452,30.712Z" />
+    <path
+        android:fillColor="#5D8F3F"
+        android:pathData="M15.849,32.831C16.005,31.869 15.352,30.962 14.39,30.806C13.428,30.65 12.521,31.303 12.365,32.265C12.209,33.227 12.862,34.134 13.824,34.29C14.786,34.446 15.693,33.793 15.849,32.831Z" />
+</vector>
