+++ core/util/src/main/res/drawable/bored.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="utf-8"?>
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="36dp"
+    android:height="36dp"
+    android:viewportWidth="36"
+    android:viewportHeight="36">
+    <path
+        android:pathData="M36,18C36,27.941 27.941,36 18,36C8.06,36 0,27.941 0,18C0,8.06 8.06,0 18,0C27.941,0 36,8.06 36,18Z"
+        android:fillColor="#FFCC4D"/>
+    <path
+        android:pathData="M25.485,27.879C25.44,27.7 24.317,23.5 18,23.5C11.682,23.5 10.56,27.7 10.515,27.879C10.46,28.096 10.558,28.321 10.752,28.433C10.947,28.544 11.191,28.511 11.352,28.356C11.371,28.337 13.306,26.5 18,26.5C22.694,26.5 24.63,28.337 24.648,28.355C24.744,28.45 24.872,28.5 25,28.5C25.084,28.5 25.169,28.479 25.246,28.436C25.442,28.324 25.54,28.097 25.485,27.879ZM29.001,14C28.696,14 28.397,13.862 28.2,13.6C25.768,10.356 21.686,12.754 21.514,12.857C21.039,13.142 20.425,12.987 20.142,12.514C19.858,12.04 20.011,11.426 20.485,11.142C22.483,9.943 26.999,8.665 29.799,12.399C30.131,12.841 30.041,13.468 29.599,13.799C29.42,13.935 29.21,14 29.001,14ZM6.999,14C6.791,14 6.58,13.935 6.4,13.8C5.958,13.469 5.869,12.842 6.2,12.4C9.001,8.666 13.517,9.944 15.514,11.143C15.988,11.427 16.141,12.041 15.857,12.515C15.573,12.988 14.961,13.143 14.487,12.859C14.308,12.753 10.213,10.384 7.799,13.601C7.604,13.862 7.303,14 6.999,14ZM29,16C29,15.448 28.553,15 28,15H21C20.447,15 20,15.448 20,16C20,16.552 20.447,17 21,17H26.092C26.299,17.581 26.848,18 27.5,18C28.328,18 29,17.329 29,16.5C29,16.39 28.986,16.283 28.964,16.179C28.976,16.119 29,16.063 29,16ZM16,16C16,15.448 15.552,15 15,15H8C7.448,15 7,15.448 7,16C7,16.552 7.448,17 8,17H13.092C13.299,17.581 13.848,18 14.5,18C15.328,18 16,17.329 16,16.5C16,16.39 15.986,16.283 15.964,16.179C15.975,16.119 16,16.063 16,16Z"
+        android:fillColor="#664500"/>
+</vector>
