+++ core/util/src/main/res/drawable/calm.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="utf-8"?>
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="36dp"
+    android:height="36dp"
+    android:viewportWidth="36"
+    android:viewportHeight="36">
+    <path
+        android:pathData="M36,18C36,27.941 27.941,36 18,36C8.059,36 0,27.941 0,18C0,8.059 8.059,0 18,0C27.941,0 36,8.059 36,18Z"
+        android:fillColor="#FFCC4D"/>
+    <path
+        android:pathData="M24.327,23.622C24.174,23.49 23.959,23.463 23.776,23.553L19.776,25.553C17.905,26.488 13.049,26.5 13,26.5C12.724,26.5 12.5,26.724 12.5,27C12.5,27.185 12.601,27.347 12.75,27.433V27.434H12.751V27.435C12.822,27.475 12.904,27.498 12.991,27.5H19.999C22.657,27.5 24.088,25.315 24.474,24.158C24.538,23.966 24.48,23.755 24.327,23.622ZM31.001,16C30.696,16 30.397,15.862 30.2,15.6C27.559,12.079 24.139,12.001 23.994,12C23.444,11.994 23,11.544 23.003,10.995C23.006,10.444 23.45,10 24,10C24.184,10 28.537,10.05 31.8,14.4C32.132,14.842 32.042,15.469 31.6,15.8C31.42,15.935 31.21,16 31.001,16ZM4.999,16C4.791,16 4.58,15.935 4.4,15.8C3.958,15.469 3.869,14.842 4.2,14.4C7.462,10.05 11.816,10 12,10C12.552,10 13,10.448 13,11C13,11.551 12.555,11.998 12.004,12C11.848,12.002 8.435,12.086 5.799,15.6C5.604,15.862 5.303,16 4.999,16ZM15.897,17.396C15.92,17.344 15.956,17.3 15.97,17.242C16.104,16.706 15.778,16.163 15.243,16.029C15.063,15.984 10.776,14.949 7.446,17.167C6.987,17.473 6.863,18.094 7.169,18.554C7.361,18.844 7.678,19 8.001,19C8.191,19 8.384,18.945 8.555,18.832C9.647,18.103 10.917,17.837 12.023,17.771C12.014,17.847 12,17.922 12,18C12,19.104 12.896,20 14,20C15.104,20 16,19.104 16,18C16,17.788 15.958,17.588 15.897,17.396ZM27.896,17.395C27.919,17.343 27.955,17.3 27.969,17.243C28.104,16.707 27.778,16.164 27.242,16.03C27.062,15.985 22.776,14.95 19.445,17.168C18.985,17.474 18.861,18.095 19.168,18.555C19.36,18.844 19.678,19 20.001,19C20.191,19 20.384,18.945 20.555,18.832C21.647,18.103 22.916,17.838 24.024,17.772C24.015,17.848 24,17.922 24,18C24,19.104 24.896,20 26,20C27.104,20 28,19.104 28,18C28,17.788 27.958,17.587 27.896,17.395Z"
+        android:fillColor="#664500"/>
+</vector>
