+++ core/util/src/main/res/drawable/depressed.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="utf-8"?>
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="36dp"
+    android:height="36dp"
+    android:viewportWidth="36"
+    android:viewportHeight="36">
+    <path
+        android:pathData="M36,18C36,27.941 27.941,36 18,36C8.06,36 0,27.941 0,18C0,8.06 8.06,0 18,0C27.941,0 36,8.06 36,18Z"
+        android:fillColor="#FFCC4D"/>
+    <path
+        android:pathData="M17.312,17.612C17.136,17.469 16.885,17.465 16.702,17.598C16.69,17.607 15.442,18.5 13,18.5C10.559,18.5 9.31,17.607 9.3,17.6C9.117,17.463 8.865,17.467 8.689,17.609C8.511,17.751 8.451,17.995 8.543,18.203C8.603,18.338 10.043,21.5 13,21.5C15.958,21.5 17.397,18.338 17.457,18.203C17.549,17.996 17.489,17.754 17.312,17.612ZM27.312,17.612C27.136,17.469 26.886,17.464 26.702,17.598C26.69,17.607 25.441,18.5 23,18.5C20.56,18.5 19.31,17.607 19.3,17.6C19.117,17.463 18.866,17.467 18.689,17.609C18.511,17.751 18.451,17.995 18.543,18.203C18.603,18.338 20.043,21.5 23,21.5C25.958,21.5 27.397,18.338 27.457,18.203C27.549,17.996 27.489,17.754 27.312,17.612ZM22,28H14C13.448,28 13,27.553 13,27C13,26.447 13.448,26 14,26H22C22.553,26 23,26.447 23,27C23,27.553 22.553,28 22,28ZM6,14C5.448,14 5,13.552 5,13C5,12.449 5.445,12.002 5.996,12C6.152,11.998 9.565,11.914 12.201,8.4C12.532,7.96 13.158,7.868 13.601,8.2C14.043,8.531 14.132,9.158 13.801,9.6C10.538,13.95 6.184,14 6,14ZM30,14C29.816,14 25.463,13.95 22.2,9.6C21.868,9.158 21.958,8.531 22.4,8.2C22.841,7.867 23.467,7.958 23.799,8.4C26.44,11.921 29.86,11.999 30.005,12C30.555,12.006 30.999,12.456 30.996,13.005C30.994,13.556 30.55,14 30,14Z"
+        android:fillColor="#664500"/>
+</vector>
