+++ core/util/src/main/res/drawable/disappointed.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="utf-8"?>
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="36dp"
+    android:height="36dp"
+    android:viewportWidth="36"
+    android:viewportHeight="36">
+    <path
+        android:pathData="M36,18C36,27.941 27.941,36 18,36C8.059,36 0,27.941 0,18C0,8.059 8.059,0 18,0C27.941,0 36,8.059 36,18Z"
+        android:fillColor="#FFCC4D"/>
+    <path
+        android:pathData="M18,21.5C8.667,21.5 7.515,29.854 7.504,29.938C7.483,30.109 7.55,30.278 7.683,30.387C7.773,30.461 7.886,30.5 8,30.5C8.053,30.5 8.106,30.491 8.158,30.475C8.217,30.455 14.126,28.5 18,28.5C23.595,28.5 27.745,30.433 27.786,30.452C27.947,30.529 28.144,30.512 28.293,30.405C28.44,30.298 28.519,30.119 28.496,29.937C28.485,29.854 27.333,21.5 18,21.5ZM29,10C23.446,10 21.198,5.633 21.105,5.447C20.858,4.953 21.058,4.352 21.552,4.105C22.045,3.859 22.644,4.057 22.892,4.548C22.967,4.694 24.713,8 29,8C29.553,8 30,8.448 30,9C30,9.552 29.553,10 29,10ZM7,10C6.448,10 6,9.552 6,9C6,8.448 6.448,8 7,8C12.083,8 12.996,4.88 13.033,4.747C13.178,4.219 13.725,3.899 14.252,4.038C14.782,4.177 15.103,4.711 14.97,5.243C14.921,5.437 13.704,10 7,10Z"
+        android:fillColor="#664500"/>
+    <path
+        android:pathData="M18,23C22,23 24,25 24,25H12C12,25 14,23 18,23Z"
+        android:fillColor="#CCD6DD"/>
+    <path
+        android:pathData="M6.001,19C5.697,19 5.397,18.862 5.2,18.6C4.868,18.159 4.958,17.532 5.4,17.201C5.543,17.094 8.351,15.018 12.256,14.268C9.781,13.027 7.034,13 6.999,13C6.447,12.998 6,12.55 6.001,11.998C6.001,11.447 6.448,11 7,11C7.221,11 12.452,11.038 15.707,14.293C15.993,14.579 16.079,15.009 15.924,15.383C15.769,15.756 15.404,16 15,16C10.387,16 6.637,18.772 6.6,18.8C6.42,18.935 6.209,19 6.001,19ZM29.999,18.999C29.791,18.999 29.581,18.935 29.401,18.801C29.363,18.772 25.59,16 21,16C20.596,16 20.23,15.756 20.076,15.383C19.921,15.009 20.007,14.579 20.293,14.293C23.549,11.038 28.779,11 29,11C29.552,11 29.998,11.447 29.999,11.998C30,12.55 29.553,12.998 29.002,13C28.966,13 26.219,13.027 23.744,14.268C27.649,15.018 30.457,17.094 30.599,17.201C31.04,17.532 31.13,18.157 30.8,18.599C30.604,18.86 30.304,18.999 29.999,18.999Z"
+        android:fillColor="#664500"/>
+</vector>
