+++ core/util/src/main/res/drawable/humorous.xml
@@ -1,0 +1,22 @@
+<?xml version="1.0" encoding="utf-8"?>
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="36dp"
+    android:height="36dp"
+    android:viewportWidth="36"
+    android:viewportHeight="36">
+    <path
+        android:pathData="M30.728,5.272C37.757,12.301 37.757,23.699 30.728,30.728C23.699,37.757 12.302,37.757 5.272,30.728C-1.757,23.699 -1.757,12.301 5.272,5.272C12.301,-1.757 23.698,-1.757 30.728,5.272Z"
+        android:fillColor="#FFCB4C"/>
+    <path
+        android:pathData="M19.279,19.28C15.785,22.774 13.06,24.684 9.634,26.994C8.853,27.523 7.706,28.923 9.635,30.853C13.492,34.71 22.746,35.1 28.922,28.924C35.099,22.747 34.71,13.494 30.851,9.636C28.922,7.706 27.521,8.854 26.993,9.636C24.685,13.059 22.774,15.785 19.279,19.28Z"
+        android:fillColor="#65471B"/>
+    <path
+        android:pathData="M11.564,28.923C11.564,28.923 15.422,26.994 21.208,21.207C26.995,15.421 28.923,11.564 28.923,11.564C28.923,11.564 30.851,17.351 24.101,24.101C17.351,30.852 11.564,28.923 11.564,28.923Z"
+        android:fillColor="#ffffff"/>
+    <path
+        android:pathData="M8.265,25.234C8.05,25.449 7.74,25.564 7.416,25.517C6.869,25.439 6.489,24.933 6.568,24.386C6.593,24.21 7.111,20.756 9.342,17.464C6.715,18.337 4.753,20.26 4.728,20.285C4.337,20.674 3.704,20.673 3.314,20.282C2.924,19.892 2.924,19.26 3.314,18.87C3.47,18.714 7.196,15.042 11.799,15.042C12.204,15.042 12.568,15.285 12.723,15.659C12.878,16.033 12.792,16.463 12.506,16.749C9.244,20.011 8.553,24.623 8.546,24.668C8.515,24.892 8.412,25.087 8.265,25.234ZM25.233,8.264C25.086,8.411 24.892,8.514 24.67,8.546C24.623,8.553 19.994,9.261 16.749,12.506C16.463,12.792 16.033,12.878 15.659,12.723C15.285,12.568 15.042,12.204 15.042,11.799C15.042,7.195 18.714,3.47 18.87,3.314C19.26,2.924 19.892,2.924 20.282,3.314C20.673,3.703 20.674,4.337 20.285,4.727C20.259,4.753 18.336,6.714 17.463,9.341C20.755,7.11 24.208,6.592 24.384,6.567C24.93,6.489 25.436,6.867 25.515,7.414C25.564,7.738 25.449,8.049 25.233,8.264Z"
+        android:fillColor="#65471B"/>
+    <path
+        android:pathData="M30.101,12.055C32.842,12.389 35.335,10.437 35.668,7.695C36.002,4.954 34.051,2.461 31.309,2.127C28.568,1.794 20.9,4.89 20.779,5.883C20.658,6.875 27.36,11.72 30.101,12.055ZM12.157,29.999C12.491,32.74 10.539,35.233 7.797,35.566C5.056,35.9 2.563,33.949 2.229,31.207C1.895,28.465 4.992,20.797 5.985,20.676C6.977,20.556 11.822,27.258 12.157,29.999Z"
+        android:fillColor="#64AADD"/>
+</vector>
