+++ core/util/src/main/res/drawable/lonely.xml
@@ -1,0 +1,37 @@
+<?xml version="1.0" encoding="utf-8"?>
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="36dp"
+    android:height="36dp"
+    android:viewportWidth="36"
+    android:viewportHeight="36">
+    <path
+        android:pathData="M18,36C27.941,36 36,27.941 36,18C36,8.059 27.941,0 18,0C8.059,0 0,8.059 0,18C0,27.941 8.059,36 18,36Z"
+        android:fillColor="#FFCC4D"/>
+    <path
+        android:pathData="M20.996,27C20.893,27 20.79,26.984 20.687,26.951C18.927,26.38 17.072,26.38 15.312,26.951C14.788,27.12 14.223,26.834 14.052,26.309C13.881,25.784 14.169,25.22 14.695,25.049C16.857,24.347 19.142,24.347 21.304,25.049C21.829,25.22 22.117,25.784 21.947,26.309C21.81,26.73 21.418,27 20.996,27Z"
+        android:fillColor="#65471B"/>
+    <path
+        android:pathData="M30.335,12.068C29.432,14.813 26.85,16.783 23.841,16.783C23.697,16.783 23.552,16.778 23.406,16.769C21.929,16.676 20.564,16.114 19.456,15.185C19.492,15.68 19.532,16.182 19.592,16.725C19.744,18.113 20.476,19.207 21.708,19.888C22.528,20.342 23.508,20.576 24.521,20.64C26.255,20.749 28.091,20.36 29.394,19.731C30.771,19.066 31.666,17.869 31.85,16.446C32.033,15.031 31.496,13.522 30.335,12.068Z"
+        android:fillColor="#ffffff"/>
+    <path
+        android:pathData="M21.351,7.583C20.054,8.133 19.404,9.884 19.374,12.872L19.413,12.94C20.31,14.259 21.786,15.164 23.501,15.272C23.615,15.279 23.729,15.283 23.842,15.283C26.476,15.283 28.691,13.346 29.095,10.759C28.98,10.654 28.874,10.547 28.752,10.443C25.037,7.273 22.285,7.186 21.351,7.583Z"
+        android:fillColor="#65471B"/>
+    <path
+        android:pathData="M23.841,16.783C26.85,16.783 29.432,14.813 30.335,12.068C29.981,11.625 29.564,11.188 29.094,10.759C28.69,13.346 26.475,15.283 23.841,15.283C23.728,15.283 23.614,15.279 23.5,15.272C21.785,15.164 20.309,14.259 19.412,12.94L19.373,12.872C19.366,13.573 19.394,14.345 19.456,15.185C20.564,16.114 21.929,16.676 23.406,16.769C23.552,16.779 23.697,16.783 23.841,16.783Z"
+        android:fillColor="#F4900C"/>
+    <path
+        android:pathData="M21.413,11.812C22.024,11.812 22.52,11.316 22.52,10.705C22.52,10.094 22.024,9.598 21.413,9.598C20.802,9.598 20.306,10.094 20.306,10.705C20.306,11.316 20.802,11.812 21.413,11.812Z"
+        android:fillColor="#ffffff"/>
+    <path
+        android:pathData="M12.159,16.783C9.15,16.783 6.568,14.813 5.665,12.068C4.504,13.522 3.968,15.031 4.15,16.445C4.335,17.868 5.229,19.066 6.606,19.73C7.909,20.359 9.744,20.748 11.479,20.639C12.492,20.575 13.472,20.342 14.292,19.887C15.523,19.206 16.255,18.112 16.408,16.724C16.468,16.182 16.508,15.682 16.544,15.188C15.441,16.111 14.074,16.675 12.594,16.768C12.448,16.779 12.303,16.783 12.159,16.783Z"
+        android:fillColor="#ffffff"/>
+    <path
+        android:pathData="M12.159,15.283C12.272,15.283 12.386,15.279 12.5,15.272C14.215,15.164 15.691,14.259 16.588,12.94L16.627,12.872C16.596,9.884 15.947,8.133 14.65,7.583C13.716,7.186 10.963,7.273 7.249,10.442C7.127,10.546 7.022,10.653 6.906,10.758C7.31,13.346 9.525,15.283 12.159,15.283Z"
+        android:fillColor="#65471B"/>
+    <path
+        android:pathData="M16.626,12.872L16.587,12.94C15.69,14.259 14.214,15.164 12.499,15.272C12.385,15.279 12.271,15.283 12.158,15.283C9.524,15.283 7.309,13.346 6.905,10.759C6.435,11.188 6.018,11.625 5.664,12.068C6.567,14.813 9.149,16.783 12.158,16.783C12.302,16.783 12.447,16.778 12.593,16.769C14.073,16.676 15.44,16.112 16.543,15.189C16.605,14.348 16.634,13.575 16.626,12.872Z"
+        android:fillColor="#F4900C"/>
+    <path
+        android:pathData="M9.781,11.81C10.391,11.772 10.855,11.246 10.816,10.636C10.778,10.026 10.252,9.562 9.642,9.6C9.032,9.638 8.568,10.164 8.606,10.774C8.645,11.384 9.171,11.848 9.781,11.81Z"
+        android:fillColor="#ffffff"/>
+</vector>
