+++ core/util/src/main/res/drawable/mysterious.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="utf-8"?>
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="34dp"
+    android:height="36dp"
+    android:viewportWidth="34"
+    android:viewportHeight="36">
+    <path
+        android:pathData="M34,17C34,26.389 26.389,34 17,34C7.611,34 0,26.389 0,17C0,7.611 7.611,0 17,0C26.389,0 34,7.611 34,17Z"
+        android:fillColor="#FFCB4C"/>
+    <path
+        android:pathData="M16.812,25.055C19.055,25.055 20.874,24.191 20.874,23.125C20.874,22.059 19.055,21.195 16.812,21.195C14.569,21.195 12.75,22.059 12.75,23.125C12.75,24.191 14.569,25.055 16.812,25.055Z"
+        android:fillColor="#65471B"/>
+    <path
+        android:pathData="M24.35,25.981L24.317,25.961C23.478,25.446 22.37,25.711 21.855,26.55L21.704,26.795C21.162,26.233 20.454,25.821 19.636,25.654L18.237,25.368L19.509,19.144C19.706,18.179 19.078,17.229 18.114,17.032L18.076,17.024C17.111,16.827 16.161,17.455 15.964,18.419L14.291,26.602L14.276,26.676L13.963,28.207L13.962,28.217L12.834,33.75C12.614,34.828 13.104,36 14.729,36H18C20.804,36 21.635,34.99 22.198,32.907L24.939,28.444C25.453,27.605 25.189,26.497 24.35,25.981Z"
+        android:fillColor="#F19020"/>
+    <path
+        android:pathData="M21.409,30.743C21.389,30.743 21.367,30.741 21.345,30.737C21.161,30.703 21.04,30.525 21.076,30.34C21.135,30.028 21.074,29.753 20.889,29.501C20.299,28.693 18.6,28.326 17.995,28.253C17.809,28.231 17.676,28.061 17.699,27.876C17.721,27.69 17.883,27.555 18.076,27.58C18.177,27.591 20.55,27.89 21.436,29.101C21.734,29.508 21.837,29.969 21.742,30.469C21.711,30.63 21.569,30.743 21.409,30.743ZM18.15,35.473C17.897,35.473 17.62,35.442 17.317,35.371C16.707,35.227 16.277,34.872 16.137,34.396C16.019,33.995 16.127,33.571 16.426,33.261C16.814,32.859 17.429,32.727 18.121,32.903C18.302,32.949 18.412,33.134 18.366,33.316C18.32,33.496 18.137,33.607 17.953,33.561C17.508,33.447 17.128,33.511 16.915,33.732C16.787,33.864 16.74,34.041 16.788,34.205C16.857,34.441 17.107,34.626 17.473,34.711C18.653,34.988 19.306,34.525 19.647,34.086C20.155,33.433 20.215,32.548 20.034,32.201C19.612,31.389 17.985,31.209 17.394,31.208C17.207,31.208 17.055,31.056 17.055,30.869C17.055,30.682 17.207,30.53 17.394,30.53C17.497,30.53 19.936,30.545 20.636,31.889C20.941,32.474 20.854,33.64 20.182,34.503C19.874,34.897 19.239,35.473 18.15,35.473ZM13.222,13.222C13.222,14.613 12.376,15.741 11.333,15.741C10.29,15.741 9.444,14.614 9.444,13.222C9.444,11.831 10.289,10.703 11.333,10.703C12.376,10.704 13.222,11.831 13.222,13.222ZM28.886,10.271C29.108,10.271 29.331,10.202 29.523,10.058C29.993,9.706 30.087,9.04 29.735,8.571C26.269,3.949 21.643,3.896 21.448,3.896C20.861,3.896 20.386,4.372 20.386,4.959C20.386,5.544 20.859,6.019 21.444,6.021C21.61,6.023 25.236,6.112 28.036,9.846C28.244,10.124 28.563,10.271 28.886,10.271ZM20.778,13.222C20.778,14.613 21.624,15.741 22.667,15.741C23.71,15.741 24.556,14.614 24.556,13.222C24.556,11.831 23.711,10.703 22.667,10.703C21.623,10.704 20.778,11.831 20.778,13.222ZM5.114,10.271C4.892,10.271 4.669,10.202 4.477,10.058C4.008,9.706 3.913,9.04 4.265,8.571C7.731,3.949 12.357,3.896 12.552,3.896C13.139,3.896 13.614,4.372 13.614,4.959C13.614,5.544 13.141,6.019 12.556,6.021C12.39,6.023 8.764,6.112 5.964,9.846C5.756,10.124 5.437,10.271 5.114,10.271Z"
+        android:fillColor="#65471B"/>
+</vector>
