+++ core/util/src/main/res/drawable/romantic.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="utf-8"?>
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="36dp"
+    android:height="36dp"
+    android:viewportWidth="36"
+    android:viewportHeight="36">
+    <path
+        android:pathData="M18,36C27.941,36 36,27.941 36,18C36,8.059 27.941,0 18,0C8.059,0 0,8.059 0,18C0,27.941 8.059,36 18,36Z"
+        android:fillColor="#FFCC4D"/>
+    <path
+        android:pathData="M16,16.958C15.581,16.958 15.191,16.693 15.051,16.274C14.848,15.675 14.033,13.958 13,13.958C11.938,13.958 11.112,15.785 10.949,16.274C10.774,16.797 10.211,17.082 9.684,16.906C9.16,16.732 8.877,16.165 9.052,15.641C9.177,15.264 10.356,11.957 13.001,11.957C15.646,11.957 16.824,15.264 16.95,15.641C17.125,16.165 16.842,16.732 16.318,16.906C16.212,16.942 16.105,16.958 16,16.958ZM26,16.958C25.581,16.958 25.191,16.693 25.052,16.274C24.849,15.675 24.033,13.958 23,13.958C21.938,13.958 21.111,15.785 20.948,16.274C20.773,16.797 20.212,17.082 19.683,16.906C19.16,16.732 18.876,16.165 19.051,15.641C19.176,15.264 20.355,11.957 22.999,11.957C25.643,11.957 26.822,15.264 26.947,15.641C27.122,16.165 26.839,16.732 26.315,16.906C26.211,16.942 26.105,16.958 26,16.958ZM27.335,20.691C27.157,20.53 26.891,20.52 26.7,20.662C26.661,20.691 22.778,23.562 18,23.562C13.234,23.562 9.338,20.691 9.3,20.662C9.109,20.52 8.843,20.532 8.665,20.691C8.488,20.851 8.448,21.115 8.571,21.319C8.7,21.534 11.788,26.562 18,26.562C24.212,26.562 27.301,21.534 27.429,21.319C27.552,21.114 27.513,20.852 27.335,20.691Z"
+        android:fillColor="#664500"/>
+    <path
+        android:pathData="M21.229,3.947C21.229,4.187 21.259,4.419 21.311,4.643C21.736,7.286 24.675,10.074 26.781,10.84C28.446,10.235 30.627,8.364 31.679,6.301C29.59,3.861 26.86,1.987 23.75,0.939C22.318,1.196 21.229,2.441 21.229,3.947ZM13.743,29.078C13.8,28.836 13.833,28.584 13.833,28.324C13.833,26.491 12.347,25.005 10.514,25.005C9.401,25.005 8.42,25.554 7.817,26.394C7.215,25.555 6.233,25.005 5.12,25.005C3.879,25.005 2.809,25.694 2.24,26.703C4.064,29.998 6.89,32.66 10.308,34.277C11.945,32.926 13.439,30.97 13.743,29.078ZM34.301,25.642C34.019,24.692 33.149,23.995 32.107,23.995C31.336,23.995 30.656,24.375 30.239,24.957C29.822,24.375 29.142,23.995 28.371,23.995C27.101,23.995 26.072,25.024 26.072,26.294C26.072,26.474 26.095,26.648 26.134,26.816C26.441,28.721 28.482,30.712 30.043,31.377C31.819,29.777 33.274,27.83 34.301,25.642Z"
+        android:fillColor="#FFAC33"/>
+    <path
+        android:pathData="M33.625,3.269C33.625,1.577 32.253,0.206 30.562,0.206C29.535,0.206 28.629,0.713 28.073,1.488C27.517,0.713 26.611,0.206 25.584,0.206C23.892,0.206 22.521,1.577 22.521,3.269C22.521,3.509 22.551,3.741 22.603,3.965C23.029,6.608 25.967,9.396 28.073,10.162C30.179,9.396 33.117,6.608 33.542,3.966C33.595,3.741 33.625,3.509 33.625,3.269ZM12.208,28.408C12.208,26.575 10.722,25.089 8.889,25.089C7.776,25.089 6.795,25.638 6.192,26.478C5.59,25.638 4.608,25.089 3.495,25.089C1.662,25.089 0.176,26.575 0.176,28.408C0.176,28.668 0.209,28.919 0.265,29.162C0.726,32.025 3.909,35.046 6.191,35.876C8.473,35.047 11.656,32.026 12.117,29.163C12.175,28.919 12.208,28.667 12.208,28.408ZM35.948,26.127C35.948,24.857 34.919,23.828 33.649,23.828C32.878,23.828 32.198,24.208 31.781,24.79C31.364,24.208 30.684,23.828 29.913,23.828C28.643,23.828 27.614,24.857 27.614,26.127C27.614,26.307 27.637,26.481 27.676,26.649C27.995,28.632 30.2,30.725 31.781,31.299C33.361,30.725 35.567,28.632 35.885,26.649C35.925,26.481 35.948,26.307 35.948,26.127Z"
+        android:fillColor="#DD2E44"/>
+</vector>
