+++ core/util/src/main/res/drawable/shameful.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="utf-8"?>
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="35dp"
+    android:height="32dp"
+    android:viewportWidth="35"
+    android:viewportHeight="32">
+    <path
+        android:pathData="M16,32C24.837,32 32,24.837 32,16C32,7.163 24.837,0 16,0C7.163,0 0,7.163 0,16C0,24.837 7.163,32 16,32Z"
+        android:fillColor="#FFCC4D"/>
+    <path
+        android:pathData="M11.471,11.529C11.471,13.235 10.484,14.617 9.265,14.617C8.046,14.617 7.059,13.235 7.059,11.529C7.059,9.824 8.046,8.441 9.265,8.441C10.483,8.441 11.471,9.824 11.471,11.529ZM23,11.529C23,13.235 22.013,14.617 20.794,14.617C19.576,14.617 18.588,13.235 18.588,11.529C18.588,9.824 19.576,8.441 20.794,8.441C22.013,8.441 23,9.824 23,11.529ZM10.118,27.382C11.883,22.97 21.589,22.97 21.589,24.735C21.588,25.618 14.529,23.853 10.118,27.382Z"
+        android:fillColor="#664500"/>
+    <path
+        android:pathData="M32.262,15.5H17C15.489,15.5 14.266,16.844 14.266,18.5C14.266,20.157 15.49,21.5 17,21.5H32.262C33.775,21.5 35,20.157 35,18.5C35,16.844 33.774,15.5 32.262,15.5Z"
+        android:fillColor="#F4900C"/>
+</vector>
