+++ core/util/src/main/res/drawable/surprised.xml
@@ -1,0 +1,25 @@
+<?xml version="1.0" encoding="utf-8"?>
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="36dp"
+    android:height="36dp"
+    android:viewportWidth="36"
+    android:viewportHeight="36">
+    <path
+        android:pathData="M36,18C36,27.941 27.941,36 18,36C8.06,36 0,27.941 0,18C0,8.06 8.06,0 18,0C27.941,0 36,8.06 36,18Z"
+        android:fillColor="#FFCC4D"/>
+    <path
+        android:pathData="M18,33C20.761,33 23,30.314 23,27C23,23.686 20.761,21 18,21C15.239,21 13,23.686 13,27C13,30.314 15.239,33 18,33Z"
+        android:fillColor="#664500"/>
+    <path
+        android:pathData="M5.999,11C5.791,11 5.58,10.935 5.4,10.8C4.958,10.469 4.869,9.842 5.2,9.4C8.462,5.05 12.816,5 13,5C13.552,5 14,5.448 14,6C14,6.551 13.555,6.998 13.004,7C12.849,7.002 9.436,7.086 6.8,10.6C6.604,10.862 6.303,11 5.999,11ZM30.001,11C29.696,11 29.397,10.862 29.2,10.6C26.56,7.079 23.139,7.002 22.994,7C22.444,6.994 22,6.544 22.003,5.995C22.006,5.444 22.45,5 23,5C23.184,5 27.537,5.05 30.8,9.4C31.132,9.842 31.042,10.469 30.6,10.8C30.42,10.935 30.21,11 30.001,11Z"
+        android:fillColor="#664500"/>
+    <path
+        android:pathData="M18,23C16.343,23 15,24.79 15,27H21C21,24.79 19.657,23 18,23Z"
+        android:fillColor="#F5F8FA"/>
+    <path
+        android:pathData="M12,18C13.381,18 14.5,16.433 14.5,14.5C14.5,12.567 13.381,11 12,11C10.619,11 9.5,12.567 9.5,14.5C9.5,16.433 10.619,18 12,18Z"
+        android:fillColor="#664500"/>
+    <path
+        android:pathData="M24,18C25.381,18 26.5,16.433 26.5,14.5C26.5,12.567 25.381,11 24,11C22.619,11 21.5,12.567 21.5,14.5C21.5,16.433 22.619,18 24,18Z"
+        android:fillColor="#664500"/>
+</vector>
