+++ core/util/src/main/res/drawable/suspicious.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="utf-8"?>
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="36dp"
+    android:height="36dp"
+    android:viewportWidth="36"
+    android:viewportHeight="36">
+    <path
+        android:pathData="M36,18C36,27.941 27.941,36 18,36C8.059,36 0,27.941 0,18C0,8.059 8.059,0 18,0C27.941,0 36,8.059 36,18Z"
+        android:fillColor="#FFCC4D"/>
+    <path
+        android:pathData="M12,19.25C13.381,19.25 14.5,18.019 14.5,16.5C14.5,14.981 13.381,13.75 12,13.75C10.619,13.75 9.5,14.981 9.5,16.5C9.5,18.019 10.619,19.25 12,19.25Z"
+        android:fillColor="#664500"/>
+    <path
+        android:pathData="M24,20C25.381,20 26.5,18.433 26.5,16.5C26.5,14.567 25.381,13 24,13C22.619,13 21.5,14.567 21.5,16.5C21.5,18.433 22.619,20 24,20Z"
+        android:fillColor="#664500"/>
+    <path
+        android:pathData="M7.102,12.559C7.351,12.735 7.676,12.795 7.987,12.694C12.174,11.339 15.014,13.248 15.133,13.33C15.586,13.642 16.208,13.531 16.523,13.08C16.838,12.629 16.731,12.009 16.282,11.692C16.132,11.586 12.546,9.117 7.372,10.791C6.846,10.961 6.558,11.525 6.729,12.05C6.797,12.264 6.932,12.438 7.102,12.559ZM25,26H11C10.448,26 10,25.553 10,25C10,24.447 10.448,24 11,24H25C25.553,24 26,24.447 26,25C26,25.553 25.553,26 25,26ZM30.001,10.408C29.633,10.408 29.27,10.247 29.022,9.937C27.627,8.186 24.723,6.65 23.708,6.65C23.066,6.65 21.783,7.561 21.104,8.218C20.608,8.7 19.816,8.685 19.337,8.19C18.857,7.695 18.868,6.905 19.361,6.424C19.752,6.044 21.797,4.149 23.708,4.149C25.713,4.149 29.246,6.205 30.978,8.378C31.409,8.918 31.319,9.704 30.78,10.134C30.549,10.319 30.273,10.408 30.001,10.408Z"
+        android:fillColor="#664500"/>
+</vector>
