+++ core/util/src/main/res/drawable/tense.xml
@@ -1,0 +1,22 @@
+<?xml version="1.0" encoding="utf-8"?>
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="36dp"
+    android:height="36dp"
+    android:viewportWidth="36"
+    android:viewportHeight="36">
+    <path
+        android:pathData="M36,18C36,27.941 27.941,36 18,36C8.06,36 0,27.941 0,18C0,8.06 8.06,0 18,0C27.941,0 36,8.06 36,18Z"
+        android:fillColor="#FFCC4D"/>
+    <path
+        android:pathData="M12,17C13.381,17 14.5,15.433 14.5,13.5C14.5,11.567 13.381,10 12,10C10.619,10 9.5,11.567 9.5,13.5C9.5,15.433 10.619,17 12,17Z"
+        android:fillColor="#664500"/>
+    <path
+        android:pathData="M24,17C25.381,17 26.5,15.433 26.5,13.5C26.5,11.567 25.381,10 24,10C22.619,10 21.5,11.567 21.5,13.5C21.5,15.433 22.619,17 24,17Z"
+        android:fillColor="#664500"/>
+    <path
+        android:pathData="M25,21C27.209,21 29,22.791 29,25C29,27.209 27.209,29 25,29H11C8.791,29 7,27.209 7,25C7,22.791 8.791,21 11,21H25Z"
+        android:fillColor="#ffffff"/>
+    <path
+        android:pathData="M25,20H11C8.243,20 6,22.243 6,25C6,27.757 8.243,30 11,30H25C27.757,30 30,27.757 30,25C30,22.243 27.757,20 25,20ZM25,22C26.483,22 27.71,23.084 27.949,24.5H24.5V22H25ZM23.5,22V24.5H20.5V22H23.5ZM19.5,22V24.5H16.5V22H19.5ZM15.5,22V24.5H12.5V22H15.5ZM11,22H11.5V24.5H8.051C8.29,23.084 9.517,22 11,22ZM11,28C9.517,28 8.29,26.916 8.051,25.5H11.5V28H11ZM12.5,28V25.5H15.5V28H12.5ZM16.5,28V25.5H19.5V28H16.5ZM20.5,28V25.5H23.5V28H20.5ZM25,28H24.5V25.5H27.949C27.71,26.916 26.483,28 25,28Z"
+        android:fillColor="#664500"/>
+</vector>
